Added visual cues for scroll start/end. need to put them in a user control
