+++ MusicPlayerForDrummers/ViewModel/MainVM.cs
@@ -32,7 +32,16 @@ namespace MusicPlayerForDrummers.ViewModel
         public PlayerVM PlayerVM { get; }
 
         private BaseViewModel? _currentViewModel;
-        public BaseViewModel? CurrentViewModel { get => _currentViewModel; set => SetField(ref _currentViewModel, value); }
+
+        public BaseViewModel? CurrentViewModel
+        {
+            get => _currentViewModel;
+            set
+            {
+                if (SetField(ref _currentViewModel, value))
+                    PlayerVM.ShowAdvancedOptions = value == PartitionVM;
+            }
+        }
 
         public DelegateCommand SwitchLibraryViewCommand { get; }
         public DelegateCommand SwitchPartitionViewCommand { get; }
