+++ MusicPlayerForDrummers/ViewModel/Sync/PageSyncVM.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Text;
+﻿using System.ComponentModel;
 
 namespace MusicPlayerForDrummers.ViewModel.Sync
 {
